requirements.txt
absl-py==1.4.0
accelerate==0.15.0
diffusers==0.11.1
nltk==3.8.1
numpy==1.24.1
torch==1.13.1
torchvision==0.14.1
transformers==4.25.1
scipy==1.10.0
tqdm==4.64.1
pandas==2.0.3
crowd-kit==1.2.1
matplotlib==3.5.2
torchmetrics==1.0.1
